+++ FrontEnd/js/api.js
@@ -18,3 +18,31 @@ export async function fetchCategory() {
   // test de fonctionnement
   // console.log(category);
 }
+
+// fonction pour filtrer les données de l'API works par catégories
+export function categoryFilter(works, categoryId) {
+  // stock toutes les catégories dans la variable "categoriesTous"
+  const categoriesTous = works;
+  // console.log("Tous", categoriesTous);
+
+  // stock toutes les catégorie dans la variable "categories1"
+  const categories1 = works.filter((work) => {
+    return work.categoryId === 1;
+  });
+  // console.log("Nombre de catégories 1 :", categories1);
+
+  // stock toutes les catégorie dans la variable "categories2"
+  const categories2 = works.filter((work) => {
+    return work.categoryId === 2;
+  });
+  // console.log("Nombre de Catégories 2 filtrer =",categories2);
+
+  // stock toutes les catégorie dans la variable "categories3"
+  const categories3 = works.filter((work) => {
+    return work.categoryId === 3;
+  });
+  // console.log("Nombre de Catégories 3 filtrer =",categories3);
+
+  // retourne les éléments ayant le même ID de catégorie que celui spécifié
+  return works.filter((work) => work.categoryId === categoryId);
+}
